+++ src/tanbirka/expressionEvaluator/api/Operations.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 interface Operatable {
-    int evaluate(int operand1, int operand2);
+    double evaluate(double operand1, double operand2);
 }
 public class Operations {
     Map<String, Operatable> operators;
@@ -14,36 +14,36 @@ public class Operations {
 
         operators.put("+", new Operatable() {
             @Override
-            public int evaluate(int operand1, int operand2) {
+            public double evaluate(double operand1, double operand2) {
                 return operand1 + operand2;
             }
         });
         operators.put("-", new Operatable() {
             @Override
-            public int evaluate(int operand1, int operand2) {
+            public double evaluate(double operand1, double operand2) {
                 return operand1 - operand2;
             }
         });
         operators.put("*", new Operatable() {
             @Override
-            public int evaluate(int operand1, int operand2) {
+            public double evaluate(double operand1, double operand2) {
                 return operand1 * operand2;
             }
         });
         operators.put("/", new Operatable() {
             @Override
-            public int evaluate(int operand1, int operand2) {
+            public double evaluate(double operand1, double operand2) {
                 return operand1 / operand2;
             }
         });
         operators.put("^", new Operatable() {
             @Override
-            public int evaluate(int operand1, int operand2) {
-                return (int) Math.pow(operand1, operand2);
+            public double evaluate(double operand1, double operand2) {
+                return (double) Math.pow(operand1, operand2);
             }
         });
     }
-    public int operate(String operator, int operand1, int operand2) {
+    public double operate(String operator, double operand1, double operand2) {
         return operators.get(operator).evaluate(operand1,operand2);
     }
 }
